ABM Model: Grid sweep All-aggregator
Final draft.

Compared with Laver and Sergenti (2011) the estimates for mean
eccentricity seems too low for higher number of firms (also the case in
the all-hunter model).

Similarly the power difference (Check 4) for mean eccentricity gives
odd results when compared with Laver and Sergenti (2011). This was also
the case for the all-hunter grid sweep model.