ABM Model: Changed total population PDF
Change the calculation of the probability density function for the
total population, such that the left subpopulation is unchanged, and
the right subpopulation is scaled according to n_ratio. [Divide by 1/4
such that the total population PDF is comparable to the PDF of the
baseline distribution, with mean (0,0) and std dev. (1,1)].