ABM model: attempted marketshare function optimisation
Changed the method used to find closest firm, from point-distance
method to one based on geometry. More precisely a bounded Voronoi
diagram is constructed of the respective firms market area and then the
inpolygon function is used to determine if customers are within a given
firms market.

However initial test show that the method did not reduce the
computational time. Instead the ‘Elapsed time’ increased from 0.279002
seconds to 1.019337 seconds when using marketshare5 compared to
marketshare4. The profile summary indicates that the inpolygon function
is partly to blame for the increase (could perhaps be reduced using
inploy mex).