ABM Model: Optimised
Small fixes that increases the computational speed.
* Changed resolution from 70 to 50. For each firm the distances is
calculated to a cell/patch in the grid. This reduces the number of
cells from 4.900 to 2.500.
* Disabled unused regionprops / regionpropsext function call.
* Changed arrayfun to (already existing) for-loop.
* Changed unnecessary pdist2 function calls to simple vectorised
distances calculations.
* Using marketshare4 rather than marketshare4 (See commit e851df3).
* Using parfor to execute the runs in parallel. (To see progress when
the program is running in parallel required changing from ‘waitbar’ to
‘parfor_progress’).

Together these changes increases the speed from approximately 320.725
ticks/hour to 8.894.185 ticks/hour (for the all-hunter model). That is
a factor 27 increase.