ABM Model: All-inductor genetic algoritm
The all-inductor model.
* Added two extra columns to the ‘cf_i’ matrix. Column 25 counts the
number of times a rule has been used to forecast position. Column 26
counts the number of times a rule has been activated.
* Fix error where original condition/forecast rules where passed to the
‘accuracy’-function. Thus accuracy was only measuring the accuracy of
the passed iteration, instead of all previous iterations. (See line
Models/ABM/ABM_ind_static.m#L294 previously #L277).
* Constructed cf_range variable for the values in the forecast rule.
This range of possible values can then be passed on to the genetic
algorithm.
* Change the default condition/forcast rule, such that its values are
always the average of all others. (Initially just the simple average,
but in the genetic algorithm the weighted average. Where the weights
are calculated based on the accuracy of each rule.)
* Added system ticks on which the genetic algorithm runs.

Genetic algorithm
* Identify worst and best performing condition/forecast rules using
fitness measure.
* Replace worst performing rules with new rules. The new rules are
constructed using either mutation or crossover.
* Mutation operation selects parent with best accuracy and randomly
flips individual bits in condition. The values of the parent is mutated
by selecting new values within the range, or by slightly modifying the
values from the parent.
* The crossover operation randomly selects bits of either parent for
the offsprings condition. The values are crossovers of the two parents.
Either component-wise crossover where each value has equal probability
to come from either parent, or weighted average of both parents values,
or the values of a randomly selected parent.
* The forecast variance/accuracy of the new condition/forecast rule is
set to the average accuracy of the two parents (if they have both been
activated), otherwise the accuracy is set to the median of all the
retained/elite rules.