Exploring alternatives to MAXCOV
Trying to develop alternative method to the current MAXCOV (where the
firm locate at centroid of “largest” delaunay triangle in delaunay
diagram without the firm, but with the boundary points).
* Two-step approach: first find the largest potential (from
intersection point of voronoi lines, intersection point of voronoi and
delaunay lines, and near-overlapping of voronoi point), and secondly
move slightly in the direction that is thought to increase market share
(using analytical results from how the area changes as firm moves along
voronoi line, or weighting the adjacent potential locations).
* Interactive voronoi/delaunay tool to help understand/look for better
approach that current MAXCOV.
* Recreated figure 11 in Averbakh et al. (2015), and used this as point
of departure.
The two-step method did not prove fruitful, since the number of points
grew exponentially as the number of firms increased.