Exploring alternatives to MAXCOV - Second attempt
Trying to develop alternative method to the current MAXCOV (where the
firm locate at centroid of “largest” delaunay triangle in delaunay
diagram without the firm, but with the boundary points).
* Based on the Voronoi cones in 2D space.
* Found key equations of consumer distribution function, using Matlab
to numerically solve the equation for different combination of mu and
n_ratio values. I.e finding peaks and saddle points of distribution.

* Well approach: The distribution is a “landscape”, and the Voronoi
cones form a “landscape”. Instead of looking for “most unsatisfied”
point/consumer (i.e largest distance to closest competitor). The
approach attempts to find the largest area of “most unsatisfied”
points/consumers. In the most simple version (with uniform
distribution), we gradually increase the height of the distribution
(think of the uniform distribution as the water level in the cone
landscape). The lowest lying areas of the cone landscape will be
submerged in water first. However it would be a point but an area (a
well, hence the name). The method compares the size of these different
“submerged” areas (wells) and chooses to located at the centroid of the
largest well. In the more elaborate version, with non-uniform
distribution, you can think of the water level as behaving strangely
(i.e. with different gravitational pull across the space). The
procedure is the same, gradually increase the height of the water
level/distribution and identify wells. Once again locate at the well
covering the largest mass of probability.

Successfully developed the well method for the uniform distribution
(although the first draft of the method was 10 times slower than the
MAXCOV delaunay triangle approach). However I was unable to get the
method to work as intended for non-uniform distributions.